+++ tests/test_consolidation.py
@@ -48,21 +48,8 @@ class Model(torch.nn.Module):
         return out
 
 
-def test_singlegpu_consolidation():
-    device = 0
-    torch.cuda.set_device(device)
-    with slapo.init_empty_weights(enable=True):
-        model = Model()
-
-    sch = slapo.create_schedule(copy.deepcopy(model))
-    sch_model, _ = slapo.build(sch, init_weights=init_module)
-
-    sch_model.cuda(device)
-    verify_weights(sch_model)
-
-
-def test_multigpu_consolidation():
-    init_dist()
+@pytest.mark.parametrize("ngpu", ["single", "multi"])
+def test_multigpu_consolidation(init_dist, ngpu):
 
     local_rank = int(os.environ["LOCAL_RANK"])
     torch.cuda.set_device(local_rank)
@@ -70,11 +57,16 @@ def test_multigpu_consolidation():
         model = Model()
 
     sch = slapo.create_schedule(copy.deepcopy(model))
-    sch["linear1"].shard("weight", axis=0)
-    sch["linear1"].shard("bias", axis=0)
-    sch["linear1"].sync(mode="bwd_post", sync_op_or_fn="all_reduce")
-    sch["linear2"].shard("weight", axis=1)
-    sch["linear2"].sync(mode="fwd_post", sync_op_or_fn="all_reduce")
+    if ngpu == "multi":
+        # Tensor parallelism.
+        sch["linear1"].shard("weight", axis=0)
+        sch["linear1"].shard("bias", axis=0)
+        sch["linear1"].sync(mode="bwd_post", sync_op_or_fn="all_reduce")
+        sch["linear2"].shard("weight", axis=1)
+        sch["linear2"].sync(mode="fwd_post", sync_op_or_fn="all_reduce")
+    else:
+        # Data parallelism.
+        pass
     sch_model, _ = slapo.build(sch, init_weights=init_module)
 
     sch_model.cuda(local_rank)
